Rearchitect to use ViewModel
Holy God, this was a pain to accomplish.

All I wanted was to be able to persist state after rotation and backgrounding. That basically amounted to spending several hours moving around all of the logic in the view that was working fine and adding a lot of confusing redirection, generally lowering the code quality. That seems to be what happens when you try to rewrite things in the way Google insists that you do. There are just so many pieces that depend on things in the view like device width and height that you have to pass around haphazardly. And there was the part where the databinding classes weren't getting autogenerated because I didn't add this one XML property that wasn't clear from the docs. Then there was another hour wondering why databinding to MutableLiveData in the layout XML wasn't working until I added adapter functions. I guess that part was just learning properly.

If ViewModel and this kind of whole program rearchitecture is the only practical way of getting in-memory persistence of large amounts of data with Android, then what were people using before MVVM was a thing? Did they just give up?